Server/requirements.txt
Flask==3.0.3
google-cloud-firestore==2.16.1
firebase-admin==6.5.0
opencv-python==4.9.0.80
dlib==19.24.1
deepface==0.0.91
tf-keras == 2.16.0
pyaudio == 0.2.14
google-cloud-speech == 2.26.1
python-dotenv
openai
requests
pdfplumber==0.5.28
flask-login==0.6.3
flask-wtf==1.2.1
flask_bootstrap==3.3.7.1
matplotlib
flask-cors==4.0.1